Module8 task1

+++ app/Http/Controllers/FavoriteController.php
@@ -5,22 +5,24 @@ namespace App\Http\Controllers;
 use App\Http\Responses\BaseResponse;
 use App\Http\Responses\FailResponse;
 use App\Http\Responses\SuccessResponse;
+use App\Models\User;
 use App\Models\Film;
 use Illuminate\Support\Facades\Auth;
+use Symfony\Component\HttpFoundation\Response;
 
 class FavoriteController extends Controller
 {
     /**
-     * Получение списка избранных фильмов
+     * Получение списка избранных фильмов.
      *
      * @return BaseResponse
      */
     public function index(): BaseResponse
     {
         try {
-            /** @var \App\Models\User|null $user */
+            /** @var User $user */
             $user = Auth::user();
-            $favoriteFilms = $user->favoriteFilms()->get();
+            $favoriteFilms = $user->favoriteFilms()->orderBy('created_at', 'desc')->get();
 
             return new SuccessResponse($favoriteFilms);
         } catch (\Exception $e) {
@@ -29,15 +31,20 @@ class FavoriteController extends Controller
     }
 
     /**
-     * Добавление фильма в избранное
+     * Добавление фильма в избранное.
      *
      * @return BaseResponse
      */
     public function store(Film $film): BaseResponse
     {
         try {
-            /** @var \App\Models\User|null $user */
+            /** @var User $user */
             $user = Auth::user();
+
+            if ($user->favoriteFilms()->where('film_id', $film->id)->exists()) {
+                return new FailResponse('Фильм уже в избранном', Response::HTTP_UNPROCESSABLE_ENTITY);
+            }
+
             $user->favoriteFilms()->attach($film);
 
             return new SuccessResponse();
@@ -47,15 +54,20 @@ class FavoriteController extends Controller
     }
 
     /**
-     * Удаление фильма из избранного
+     * Удаление фильма из избранного.
      *
      * @return BaseResponse
      */
     public function destroy(Film $film): BaseResponse
     {
         try {
-            /** @var \App\Models\User|null $user */
+            /** @var User $user */
             $user = Auth::user();
+
+            if (!$user->favoriteFilms()->where('film_id', $film->id)->exists()) {
+                return new FailResponse('Фильм не найден в избранном', Response::HTTP_UNPROCESSABLE_ENTITY);
+            }
+
             $user->favoriteFilms()->detach($film);
 
             return new SuccessResponse();
